# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SNU Archive API
  version: 2.0.0-draft
  summary: Exam score distributions and difficulty ratings for SNU courses.
  description: |
    Contract between the Go backend and the React frontend. Revised from the
    first draft (v1.0.0) to match `docs/superpowers/specs/2026-09-27-go-backend-design.md`.

    **Authentication.** Google OAuth restricted to `@snu.ac.kr`. On success the
    server sets `snu_session` (HttpOnly, signed) and `snu_csrf` (readable).
    Every unsafe method must echo the CSRF cookie in `X-CSRF-Token`, and its
    `Origin` header must equal the configured app origin. OpenAPI cannot
    express "header must equal cookie"; that pairing is a server-side check.

    **Same origin.** The API is served under the app's own host (`/api/v1`),
    so no CORS is configured and `SameSite=Lax` stays viable.

    **Semesters** are `1` spring (1학기), `2` summer (여름학기), `3` fall
    (2학기), `4` winter (겨울학기). Ordering by `(year, semester)` is calendar
    order. Labels come from `/config`; clients never derive them.

    **Exam sittings.** Statistics, votes and voting requests all attach to an
    exam sitting: one course × assessment kind × number × term. Sittings are
    created on demand by the server (get-or-create) whenever something first
    refers to them.

    **Pagination.** Keyset cursors. `cursor` is opaque; `nextCursor` is null on
    the last page. `limit` defaults to 20, maximum 50.

...
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /auth/dev-login:
    post:
      operationId: devLogin
      tags: [dev]
      summary: Sign in without Google (development only)
      description: |
        Exists only when `APP_ENV=development` and `DEV_LOGIN_ENABLED=true`;
        otherwise the route is not registered and returns 404. The server
        refuses to start with this enabled in production.

        Upserts the user exactly as the OAuth callback would, including the
        `ADMIN_EMAILS` check, then sets both cookies.

        Checks the `Origin` header (it must equal the app origin; otherwise
        `403 CSRF_INVALID`) but needs no CSRF token, since no session exists
        yet. `/config.devLoginEnabled` tells the client whether this exists.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  description: Must end in `@snu.ac.kr`.
                displayName: { type: [string, 'null'], maxLength: 120 }
              additionalProperties: false
      responses:
        '204': { description: Signed in }
        '400': { $ref: '#/components/responses/Malformed' }
        '403': { $ref: '#/components/responses/CsrfInvalid' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: snu_session
      description: |
        HttpOnly, Secure, SameSite=Lax. HMAC-signed `{uid, exp, ep}`; the
        server also checks the account is live and `ep` equals its session
        epoch.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Must equal the readable `snu_csrf` cookie; mismatch is
        `403 CSRF_INVALID`. The `Origin` header must also match the app origin.
    cronSecret:
      type: http
      scheme: bearer
      description: The `CRON_SECRET` environment value.
...
  responses:
    Malformed:
      description: Unparseable body, bad cursor or bad query parameter (`MALFORMED_REQUEST`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    CsrfInvalid:
      description: CSRF token or Origin check failed (`CSRF_INVALID`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Unknown id or route (`NOT_FOUND`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationFailed:
      description: Field-level validation failure (`VALIDATION_FAILED`, `details.fields`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message:
              type: string
              description: Korean fallback for logs and unhandled cases. Not UI copy.
            requestId:
              type: string
              description: Matches the `X-Request-ID` response header and server logs.
            details:
              type: object
              additionalProperties: true
              properties:
                fields:
                  type: array
                  items: { $ref: '#/components/schemas/FieldError' }
              description: |
                Code-specific. `VALIDATION_FAILED` carries
                `fields: FieldError[]`; `EXPORT_TOO_LARGE` carries `count` and
                `limit`; `FILE_TOO_LARGE` carries `limit`.
    ErrorCode:
      type: string
      enum:
        - MALFORMED_REQUEST
        - NOT_AUTHENTICATED
        - CSRF_INVALID
        - ADMIN_REQUIRED
        - NOT_FOUND
        - METHOD_NOT_ALLOWED
        - VALIDATION_FAILED
        - CONFIRMATION_REQUIRED
        - VOTING_ALREADY_OPEN
        - VOTING_NOT_OPEN
        - VOTING_REQUEST_EXISTS
        - VOTING_REQUEST_NOT_OPEN
        - NOT_REQUEST_OWNER
        - REPORT_ALREADY_REVIEWED
        - FILE_TOO_LARGE
        - FILE_TYPE_REJECTED
        - LAST_ADMIN_PROTECTED
        - ENV_ADMIN_PROTECTED
        - EXPORT_TOO_LARGE
        - DELETE_PREVIEW_MISMATCH
        - JOB_DISABLED
        - JOB_ALREADY_RUNNING
        - INTERNAL
    FieldError:
      type: object
      required: [field, code]
      properties:
        field:
          type: string
          description: |
            JSON pointer-ish path of the offending field, e.g. `q3`. `""` means
            the error concerns the body as a whole (e.g. `NOTHING_SUBMITTED`)
            or several fields at once (`QUARTILES_OUT_OF_ORDER`, and
            `VALUE_OUT_OF_RANGE` on statistics).
        code:
          type: string
          description: Named after the database constraint that would have caught it.
          enum:
            - REQUIRED
            - TOO_LONG
            - QUARTILES_OUT_OF_ORDER      # stat_reports_ordered_ck
            - VALUE_ABOVE_MAX_SCORE       # stat_reports_within_max_ck
            - VALUE_OUT_OF_RANGE          # stat_reports_range_ck
            - NOTHING_SUBMITTED           # stat_reports_not_empty_ck
            - UNKNOWN_ASSESSMENT_KIND
            - INVALID_ASSESSMENT_NUMBER   # exam_sittings number trigger
            - INVALID_TERM
            - INVALID_COLLEGE
            - INVALID_ADMISSION_YEAR
            - INVALID_EMAIL
            - CLOSES_AT_IN_PAST
            - WINDOW_INVERTED
            - INVALID_FAVORITE_ORDER      # ids is not exactly the caller's favourite set

    # ------------------------------------------------------ reference data
...